Update README.md
I tried to get on libertree, and was given an error message, so I tried to get around that by getting a message to pistos on maple via rekado on oak and ended up here...so now I'll see what this is about and maybe get back to looking for rekado at that time. My name on libertree is marsee, btw.